refactor: extract instace type to instance file
